requirements.txt
createnv==0.0.1
dj-database-url==0.5.0
django==4.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
python-telegram-bot==13.10
